refactor: add folder utils and put there utils
