+++ src/commands/cat.js
@@ -1,6 +1,6 @@
 import fs from 'fs';
-import getFullPath from './getFullPath.js';
-import showDirectory from './showDirectory.js';
+import getFullPath from '../utils/getFullPath.js';
+import showDirectory from '../utils/showDirectory.js';
 
 async function executeCat(pathToFile) {
     const filePath = await getFullPath(pathToFile);
