+++ src/commands/cd.js
@@ -1,7 +1,7 @@
 import * as fs from 'fs/promises';
 import { currentDirectory } from '../file-manager.mjs';
-import getFullPath from './getFullPath.js';
-import showDirectory from './showDirectory.js';
+import getFullPath from '../utils/getFullPath.js';
+import showDirectory from '../utils/showDirectory.js';
 
 async function executeCd(pathToFolder){
     // Go to dedicated folder from current directory
