+++ src/commands/decompress.js
@@ -2,9 +2,9 @@ import * as fsPromise from 'fs/promises';
 import fs from 'fs';
 import zlib from 'zlib';
 import { pipeline } from 'stream';
-import getFullPath from './getFullPath.js';
-import showDirectory from './showDirectory.js';
-import splitTwoArgs from './splitTwoArgs.js';
+import getFullPath from '../utils/getFullPath.js';
+import showDirectory from '../utils/showDirectory.js';
+import splitTwoArgs from '../utils/splitTwoArgs.js';
 
 async function decompress(arg) {
