+++ src/commands/ls.js
@@ -1,6 +1,6 @@
 import * as fs from 'fs/promises';
 import { currentDirectory } from '../file-manager.mjs';
-import showDirectory from './showDirectory.js';
+import showDirectory from '../utils/showDirectory.js';
 
 async function executeLs(){
     // List all files and folder in current directory and print it to console
