+++ src/commands/remove.js
@@ -1,7 +1,6 @@
 import * as fs from 'fs/promises';
-import showDirectory from './showDirectory.js';
-import path from 'path';
-import getFullPath from './getFullPath.js';
+import showDirectory from '../utils/showDirectory.js';
+import getFullPath from '../utils/getFullPath.js';
 
 async function executeRm(pathToFile){
     const fullPath = await getFullPath(pathToFile);
