+++ src/commands/up.js
@@ -1,6 +1,6 @@
 import { currentDirectory } from '../file-manager.mjs';
 import path from 'path';
-import showDirectory from './showDirectory.js';
+import showDirectory from '../utils/showDirectory.js';
 
 async function executeUp(){
     // Go upper from current directory (when you are in the root folder
